+++ tests/test_requests.py
@@ -326,29 +326,35 @@ def test_request_send_push_promise():
         await request.send_push_promise("/style.css")
         response = JSONResponse({"json": "OK"})
         await response(scope, receive, send)
+
     client = TestClient(app)
     response = client.get("/")
     assert response.json() == {"json": "OK"}
+
 
 def test_request_send_push_promise_without_push_extension():
     """
     If server does not support the `http.response.push` extension,
     .send_push_promise() does nothing.
     """
+
     async def app(scope, receive, send):
         request = Request(scope)
         await request.send_push_promise("/style.css")
         response = JSONResponse({"json": "OK"})
         await response(scope, receive, send)
+
     client = TestClient(app)
     response = client.get("/")
     assert response.json() == {"json": "OK"}
+
 
 def test_request_send_push_promise_without_setting_send():
     """
     If Request is instantiated without the send channel, then
     .send_push_promise() is not available.
     """
+
     async def app(scope, receive, send):
         # the server is push-enabled
         scope["extensions"]["http.response.push"] = {}
@@ -360,6 +366,7 @@ def test_request_send_push_promise_without_setting_send():
             data = "Send channel not available"
         response = JSONResponse({"json": data})
         await response(scope, receive, send)
+
     client = TestClient(app)
     response = client.get("/")
     assert response.json() == {"json": "Send channel not available"}