+++ changelog.md
@@ -3,6 +3,7 @@
 2. StaticFiles: dir => dir/index.html, not found(html=True): 404.html
 3. Adding a percent sign to redirect with quoted params
 4. app run: trace-id
+5. server-push , app state
 
 # v0.2.5
 1. Opt: Add ImmutableMultiDict, Auth: Request -> HttpConnection, graphql: executor_class